Make BLAST protein domain database GUI consistent.
Not adding a new datatype yet, but plan to do so in future. Use the same
style options as the BLAST protein/nucleotide database tools so that any
jobs with the current RPS-BLAST or RPS-TBLASTN tools will still work in
future when they also support a BLAST protein domain database in the
user's history.