TravisCI integration for continous integration via Galaxy tool functional tests.
This works by installing a copy of Galaxy and the BLAST+ binaries,
doing a 'manual install' of the Galaxy Tools in this repository,
and then running the Galaxy tool functional test script.

The installation including the test suite setting up the database
(via the migration scripts) alone takes about 3m30s. This is not
ideal... but running the tests takes quite a bit longer :(

To minimise the test run time, we use a custom tool_conf.xml with
only our tools (and the data upload tool) to avoid wasting time
loading tools we will not be testing. Similarly we use a custom
datatypes_conf.xml with just the datatypes of interest to avoid
wasting time loading the other formats.

We are also using wget from a GitHub mirror of the Galaxy repository,
git clone of the mirror was a little slower, wget from the official
Galaxy bitbucket slower still, and an hg close slowest of all.