Test Blast2GO under TravisCI.
This requires creating configuration files using sed (as done in
the tool_dependencies.xml file for Blast2GO), and also needs Java.

The simpliest solution to getting Java on the TravisVM was to run
the entire test suite as a Java job since the default Java image
includes a suitable version of Python 2 as well.