Support the new taxonomy columns in pick-your-own tabular output.
NOTE: This assumes BLAST+ will be able to find the taxdb.* files
via the BLASTDB environment variable or otherwise. If not, then
the taxonomy text files will be just N/A.

NOTE: This does not address how to track/support different versions
of the NCBI taxonomy (which does change over time). See issue #21

NOTE: We can't yet include this in the unit tests as it needs
a BLAST database as a test input. See issue #3