Include pident and qcovhsp values in blast_rbh output
Is it worth showing just the minimum score (A vs B and B vs A),
since for the most part the bitscore and percentage identify
will be the same (often get the exact same alignment)?

Note that the query coverage number does vary in cases where one
gene is longer than the other. That is potentially interesting.

Scope for adding other fields, like the descriptions and perhaps
sequence lengths?