Update blast_rbh.py
I was looking into the code for reciprocal best blast with a similar code that I already have and I notice that some best hits where missing with this code. score > best_score doesn't allow for the maximum score to be the first find. 
so I needed to add the equals >= and now it works, those hits didn't have any ties.