Revert "TravisCI: Run linting as initial TravisCI stage"
This reverts commit 37141230e1ac8e46b5e79d5d7c58511649664e52.

This worked, but the time spent linting (about 10s) is dwarfed by
the total overhead of a TravisCI build (over a minute, including
up to 30s just to install planemo and its dependencies).

The time taken to install flake8 and restructuredtext-lint is
minor.

Overall it is better to do the linting at the start of each
test build (which already have to install planemo anyway).